deleted files to consolidate into Schedule.cs
